Updated requirements.txt to include my new package that should fix betterforms to work properly in 2022.
Forked django-betterforms, then updated and packaged for install at...
pip install git+https://github.com/chrisdotgrubb/django-betterforms

Betterforms is great, but hasn't been updated in a few years. Still has python2 parts, and old locations of some functions. I updated betterforms (specifically the MultiForm and MultiModelForm) to be used without issues in python3 and django 4.0 and above. That package should work in >=django 3.0 also, but haven't tested that. The normal forms, views, and tests in betterforms may or may not work. But all that I use is the MultiForm, which does work. I may go back and fix the others later if necessary, but for now it is working for what I need.